Split up sql dialects by consuming store
