+++ src/ServiceControl.Persistence.EFCore/Implementation/RetryBatchStore.cs
@@ -8,7 +8,7 @@ using ServiceControl.Persistence.EFCore.Entities;
 using ServiceControl.Persistence.EFCore.Infrastructure;
 using ServiceControl.Persistence.Infrastructure;
 
-public class RetryBatchStore(IServiceScopeFactory scopeFactory, IIngestionSqlDialect dialect) : DataStoreBase(scopeFactory), IRetryBatchStore
+public class RetryBatchStore(IServiceScopeFactory scopeFactory, IRetryBatchSqlDialect dialect) : DataStoreBase(scopeFactory), IRetryBatchStore
 {
     public Task<string> CreateBatch(string retrySessionId, string requestId, RetryType retryType,
         string[] failedMessageRetryIds, string originator, DateTime startTime, DateTime? last = null,
